requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
azure-storage-blob==12.19.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1